fix: Prevent active session API call being made twice (#266)
Refactor RemoteSessionReader to ensure it calls the session API only as many times as needed.

Initially, the 'prove your identity' card checks if the user has an active ID Check session before displaying. However, due to the way it was listening for config changes, this API call was unnecessarily firing twice to start with.