refactor: Simplify session reader from `Flow` to `suspend` function (#291)
Simplify `SessionReader.isActiveSession()` from `Flow` to `suspend` function.

DCMAW-12575

We made the `RemoteSessionReader` reactive to changes in config to accommodate the test wrapper setup where config could be changed after launching the journey.

This now causes problems as, in order to continue smoothly after process death, we need to resubscribe to the Flow and hence retrigger the API calls. If the API call fails, we clear the session from the store, which is not what we want.

With the new test wrapper setup this reactive session reader is no longer needed.