feat: Apply default config on SDK initialisation (#299)
- Apply default config on SDK initialisation
- Validate that all config is provided

Currently consuming apps need to provide all the configuration, regardless of whether they want to customise a given configuration value. This means that each new configuration key that we require becomes a breaking change. As the project adds more configuration options, consuming apps need to keep adding more config to avoid a breaking change, which is not scalable.

In DCMAW-11490, we'll need to add another configuration parameter to help with testing the ID Check SDK: whether to enable the manual launcher (see #298). This change enables us to add a new config option in a non-breaking way.